+++ utils/common.ts
@@ -1,7 +1,11 @@
 import { ref } from 'vue'
+import { onShareAppMessage } from '@dcloudio/uni-app'
+import { store } from '@/uni_modules/uni-id-pages/common/store.js'
+import { omitBy, isNil } from 'lodash'
 
+const db = uniCloud.database()
 
-export function tipFactory(storage: string, showBool: ref<string>, closeFunctionName: ref<{func:string}>) {
+export function tipFactory(storage: string, showBool: ref<string>, closeFunctionName: ref<{ func: string }>) {
     return function () {
         return new Promise((resolve) => {
             if (!uni.getStorageSync(storage)) {
@@ -18,5 +22,55 @@ export function tipFactory(storage: string, showBool: ref<string>, closeFunction
                 resolve()
             }
         })
+    }
+}
+
+export function shareMessageCall(res) {
+    console.log(res)
+    const pages = getCurrentPages()
+    const currentPage = pages[pages.length - 1]
+    const currentPath = currentPage.route
+    return {
+        title: '是时光丫',
+        path: `${currentPath}?inviteCode=${store.userInfo.my_invite_code}&sceneId=onShareAppMessage_${+new Date()}`,
+    }
+}
+
+export function saveSceneId(sceneDetails) {
+    const { inviteCode, _id, sceneId } = sceneDetails
+    //如果导入用户分享的二维码时，二维码中的用户id与自身的邀请用户id一致，且inviter_scene_id为空
+    //则视为该用户为该二维码引流的新用户，将二维码id写入当前用户信息中，以便后期分析用户来源
+    //采取逻辑，则无需要uni-id-page中注册逻辑实现
+    if (store.userInfo.inviteCode && store.userInfo.inviteCode === inviteCode && !store.userInfo.inviter_scene_id) {
+        const params = {
+            inviter_special_day_id: _id, //用于后续统计分享日期数据
+            inviter_scene_id: sceneId, //分享来源记录（海报id,群聊链接id,朋友圈id）
+        }
+
+        db.collection('uni-id-users').where("'_id' == $cloudEnv_uid").update(omitBy(params, isNil))
+        //发放给邀请人
+        inviterAward(store.userInfo.inviter_uid[0])
+    }
+}
+async function inviterAward(userId) {
+    const uniScores = db.collection('uni-id-scores')
+    try {
+        const res = await uniScores
+            .where({
+                user_id: userId,
+            })
+            .field('balance')
+            .orderBy('create_date desc')
+            .limit(1)
+            .get()
+        await uniScores.add({
+            user_id: userId,
+            balance: res?.result?.data[0].balance + 5,
+            score: 5,
+            type: 1,
+            comment: '邀请新用户获得',
+        })
+    } catch (e) {
+        console.log(e)
     }
 }